move is_function
