+++ qcodes/utils/__init__.py
@@ -3,6 +3,7 @@ from .abstractmethod import qcodes_abstractmethod
 from .attribute_helpers import DelegateAttributes, checked_getattr, strip_attrs
 from .delaykeyboardinterrupt import DelayedKeyboardInterrupt
 from .deprecate import QCoDeSDeprecationWarning, deprecate, issue_deprecation_warning
+from .function_helpers import is_function
 from .installation_info import (
     convert_legacy_version_to_supported_version,
     get_all_installed_package_versions,
@@ -28,4 +29,5 @@ __all__ = [
     "issue_deprecation_warning",
     "create_on_off_val_mapping",
     "qcodes_abstractmethod",
+    "is_function",
 ]
