+++ qcodes/utils/helpers.py
@@ -38,6 +38,7 @@ from ruamel.yaml import YAML
 
 from .abstractmethod import qcodes_abstractmethod as abstractmethod
 from .attribute_helpers import DelegateAttributes, checked_getattr, strip_attrs
+from .function_helpers import is_function
 from .json_utils import NumpyJSONEncoder
 from .path_helpers import QCODES_USER_PATH_ENV, get_qcodes_path, get_qcodes_user_path
 from .val_mapping import create_on_off_val_mapping
@@ -54,48 +55,6 @@ if TYPE_CHECKING:
 
 
 log = logging.getLogger(__name__)
-
-
-def is_function(f: object, arg_count: int, coroutine: bool = False) -> bool:
-    """
-    Check and require a function that can accept the specified number of
-    positional arguments, which either is or is not a coroutine
-    type casting "functions" are allowed, but only in the 1-argument form.
-
-    Args:
-        f: Function to check.
-        arg_count: Number of argument f should accept.
-        coroutine: Is a coroutine.
-
-    Return:
-        bool: is function and accepts the specified number of arguments.
-
-    """
-    if not isinstance(arg_count, int) or arg_count < 0:
-        raise TypeError('arg_count must be a non-negative integer')
-
-    if not (callable(f) and bool(coroutine) is iscoroutinefunction(f)):
-        return False
-
-    if isinstance(f, type):
-        # for type casting functions, eg int, str, float
-        # only support the one-parameter form of these,
-        # otherwise the user should make an explicit function.
-        return arg_count == 1
-
-    try:
-        sig = signature(f)
-    except ValueError:
-        # some built-in functions/methods don't describe themselves to inspect
-        # we already know it's a callable and coroutine is correct.
-        return True
-
-    try:
-        inputs = [0] * arg_count
-        sig.bind(*inputs)
-        return True
-    except TypeError:
-        return False
 
 
 def full_class(obj: object) -> str:
